Move map config to config.apps.map
