+++ src/config/aachen/index.ts
@@ -1,10 +1,10 @@
-import map from './map';
 import menu from './menu';
 import routes from './routes';
 import staticpages from './staticpages';
 
 import defaultColors from '~/config/default/colors';
 import defaultBase from '~/config/default/base';
+import { map as defaultMapConfig } from '~/config/default/apps';
 import { RegionConfig } from '~/types';
 
 const apiEndpoints = {
@@ -14,7 +14,12 @@ const apiEndpoints = {
 };
 
 const aachenConfig: RegionConfig = {
-  map,
+  apps: {
+    map: {
+      ...defaultMapConfig,
+      geocoderBounds: '5.974,50.648,6.217,50.858'
+    }
+  },
   menu,
   routes,
   staticpages,
