--- src/config/aachen/map.ts
@@ -1,6 +1,0 @@
-import defaultConfig from '../default/map';
-
-export default {
-  ...defaultConfig,
-  geocoderBounds: '5.974,50.648,6.217,50.858'
-};
