+++ src/config/default/index.ts
@@ -1,14 +1,16 @@
 import { DefaultConfig } from '~/types';
+import { map } from './apps';
 import baseConfig from './base';
 import colors from './colors';
-import map from './map';
 import menu from './menu';
 import staticpages from './staticpages';
 import routes from './routes';
 
 const defaultConfig: DefaultConfig = {
+  apps: {
+    map
+  },
   colors,
-  map,
   menu,
   routes,
   staticpages,
